Moved ChartComponent into dashboard-plugin module.
